Move debug configuration to API
